Define doctest helper in docs/conftest.py
